refactor: Moving interfaces folder to common folder
