+++ src/lib/common/interfaces/AoeffModel.ts
