+++ src/lib/common/interfaces/ArchiModel.ts
