+++ src/lib/common/interfaces/model/Node.ts
