+++ src/lib/common/interfaces/model/index.ts
@@ -1,0 +1,29 @@
+import { Model, Nodes, TotalByType, Relationships } from '@lib/common/interfaces/model/Model';
+import {
+  View,
+  ViewNode,
+  ViewRelationship,
+  BendPoint,
+  Bounds,
+} from '@lib/common/interfaces/model/View';
+import { Statistics } from '@lib/common/interfaces/model/Statistics';
+import { Node, Property } from '@lib/common/interfaces/model/Node';
+import { Relationship } from '@lib/common/interfaces/model/Relationship';
+import { Landscape } from '@lib/common/interfaces/model/Landscape';
+
+export {
+  Model,
+  Nodes,
+  TotalByType,
+  Relationships,
+  View,
+  ViewNode,
+  ViewRelationship,
+  BendPoint,
+  Bounds,
+  Statistics,
+  Landscape,
+  Relationship,
+  Node,
+  Property,
+};
