--- src/lib/interfaces/model/index.ts
@@ -1,23 +1,0 @@
-import { Model, Nodes, TotalByType, Relationships } from '@lib/interfaces/model/Model';
-import { View, ViewNode, ViewRelationship, BendPoint, Bounds } from '@lib/interfaces/model/View';
-import { Statistics } from '@lib/interfaces/model/Statistics';
-import { Node, Property } from '@lib/interfaces/model/Node';
-import { Relationship } from '@lib/interfaces/model/Relationship';
-import { Landscape } from '@lib/interfaces/model/Landscape';
-
-export {
-  Model,
-  Nodes,
-  TotalByType,
-  Relationships,
-  View,
-  ViewNode,
-  ViewRelationship,
-  BendPoint,
-  Bounds,
-  Statistics,
-  Landscape,
-  Relationship,
-  Node,
-  Property,
-};
